+++ src/player/Player.java
@@ -29,7 +29,7 @@ public class Player extends Actor {
     private Vector2 absPos;
     private Weapon weapon;
 
-    public Player(String name, int id, int x, int y, boolean canMove) {
+    public Player(String name, int id, double x, double y, boolean canMove) {
         this.name = name;
         this.id = id;
         this.canMove = canMove;
@@ -45,7 +45,7 @@ public class Player extends Actor {
         weapon = new Pistol();
     }
 
-    public Player(String name, int id, int x, int y) {
+    public Player(String name, int id, double x, double y) {
         this(name, id, x, y, true);
     }
 
@@ -69,15 +69,15 @@ public class Player extends Actor {
         this.id = id;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.setPosition(x, this.getY());
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.setPosition(this.getX(), y);
     }
 
-    public void move(int dx, int dy) {
+    public void move(double dx, double dy) {
         setX(getX() + dx);
         setY(getY() + dy);
     }
@@ -89,8 +89,8 @@ public class Player extends Actor {
                 keyListener = getKeyboard();
                 mousePos = MouseInfo.getPointerInfo();
                 Point b = mousePos.getLocation();
-                int gunX = (int) b.getX();
-                int gunY = (int) b.getY();
+                double gunX = b.getX();
+                double gunY = b.getY();
                 if (keyListener.isKeyPressed("W")) {
                     //weapon.move(1,"NORTH");
                     move(1, "NORTH");
@@ -133,11 +133,11 @@ public class Player extends Actor {
         return tag;
     }
 
-    public int getAbsX() {
+    public double getAbsX() {
         return absPos.getX();
     }
 
-    public int getAbsY() {
+    public double getAbsY() {
         return absPos.getY();
     }
 
@@ -145,7 +145,7 @@ public class Player extends Actor {
         return absPos;
     }
 
-    public void setAbsPos(int x, int y) {
+    public void setAbsPos(double x, double y) {
         absPos.set(x, y);
         weapon.setAbsPos(x, y);
     }
