+++ src/server/ServerManager.java
@@ -35,8 +35,8 @@ public class ServerManager extends Server {
                 }
             }
         } else if (command[0].equals("move")) {
-            int xNew = Integer.parseInt(command[1]);
-            int yNew = Integer.parseInt(command[2]);
+            double xNew = Double.parseDouble(command[1]);
+            double yNew = Double.parseDouble(command[2]);
 
             for (Player player : game.getPlayers()) {
                 if (player.getId() == i) {
@@ -47,10 +47,10 @@ public class ServerManager extends Server {
             }
         } else if (command[0].equals("shoot")) {
             for (Player player : game.getPlayers()) {
-                int x = Integer.parseInt(command[1]);
-                int y = Integer.parseInt(command[2]);
-                int vx = Integer.parseInt(command[3]);
-                int vy = Integer.parseInt(command[4]);
+                double x = Double.parseDouble(command[1]);
+                double y = Double.parseDouble(command[2]);
+                double vx = Double.parseDouble(command[3]);
+                double vy = Double.parseDouble(command[4]);
                 send(player.getId(), "shoot " + x + " " + y + " " + vx + " " + vy);
             }
         }
